[VL 2.0.0] Feature: Implement Unified Number Type with Floating Point Support

+++ ast/ast.go
@@ -208,7 +208,7 @@ type Identifier struct {
 func (i *Identifier) GetType() NodeType { return IdentifierNode }
 
 type NumericLiteral struct {
-	Value int // TODO: Switch to float64
+	Value float64
 }
 
 func (n *NumericLiteral) GetType() NodeType { return NumericLiteralNode }
